+++ HISTORY.rst
@@ -6,6 +6,13 @@ Release History
 dev
 +++
 
+**Improvements**
+- Support for editing markdown files in Jupyter
+- New pre-save hook `update_selected_formats` that saves to formats in metadata 'nbrmd_formats'
+
+**BugFixes**
+- ContentManager compatible with Python 2.7
+
 0.2.0 (2018-06-21)
 +++++++++++++++++++